requirements.txt
streamlit
Pillow
natsort